Kindle: alternate certificate chains, and two fixes from the HiBreak port
TLS reaches the handshake and completes it: the device log carries
"handshake ok (auto): TLSv1.3 / TLS_AES_128_GCM_SHA256 in 467 ms".

wolfSSL gains --enable-altcertchains. By default it requires the chain a
server PRESENTS to end at a trusted root, and cross-signed chains do not:
www.gutenberg.org sends leaf, Network Solutions CA 3, USERTrust
cross-signed, and finally AAA Certificate Services, which is self-signed
and no longer in Mozilla's bundle. The trusted anchor sits in the MIDDLE
of what was sent. Alternative chains let any presented certificate that
is directly trusted end the walk, which is what browsers and OpenSSL have
always done, and without it the failure is ASN_NO_SIGNER_E — which reads
as "the root is missing" while the root is loaded.

Two fixes handed over from the HiBreak port, verified there:

DirectPixelWriter indexed the framebuffer with a uint16_t, so the offset
wrapped on any panel past 524,288 pixels. This one is 600x800, 60,000
bytes, and fits with 5,535 to spare — by luck. It wrapped on the 824x1648
HiBreak, and since the stride does not divide 65,536 the wraps scattered
rather than repeated, so a page looked right except for its illustration.
Upstream's bug rather than a port's, and worth sending there.

logPrintf gains __attribute__((format(printf, 3, 4))). Without it not one
LOG_ call site in the tree is checked. It only bites where
ENABLE_SERIAL_LOG is defined, which this build does not, so the warnings
it would raise are invisible here — and that is itself worth knowing,
since this build may be carrying call sites that do not compile.

+++ lib/Epub/Epub/converters/DirectPixelWriter.h
@@ -179,7 +179,15 @@ struct DirectPixelWriter {
     const int sy = phyY - originY;
     if (static_cast<unsigned>(sy) >= static_cast<unsigned>(clipRows)) return;
 
-    const uint16_t byteIndex = static_cast<uint16_t>(sy * displayWidthBytes + (phyX >> 3));
+    // uint32_t, not uint16_t. The index is a byte offset into the whole frame,
+    // so it exceeds 65,535 on any panel past 524,288 pixels. This one is
+    // 600x800, which is 60,000 bytes and fits with 5,535 to spare — by luck,
+    // not by design. It wrapped on the 824x1648 HiBreak, where 169,744 bytes
+    // meant every row past 636 landed at a wrong offset, and since the stride
+    // does not divide 65,536 the wraps scattered rather than repeated. Only
+    // images come through this writer, so a page looked right except for its
+    // illustration. Found in the HiBreak port and upstream's bug, not a port's.
+    const uint32_t byteIndex = static_cast<uint32_t>(sy) * displayWidthBytes + static_cast<uint32_t>(phyX >> 3);
     const uint8_t bitMask = 1 << (7 - (phyX & 7));
 
     if (state) {
